+++ inst/include/single_sample.h
@@ -6,7 +6,7 @@
 
 namespace spacefillr {
 
-static inline std::vector<float> sobol_fast_calc_std(unsigned long long  i, unsigned int dim, unsigned int scramble) {
+static inline std::vector<float> sobol_fast_calc_std(uint64_t  i, unsigned int dim, unsigned int scramble) {
   std::vector<float> vals(dim);
   for(unsigned int j = 0; j < dim; j++) {
     vals[j] = spacefillr::sobol_owen_fast(i, j, scramble);
@@ -14,7 +14,7 @@ static inline std::vector<float> sobol_fast_calc_std(unsigned long long  i, unsi
   return(vals);
 }
 
-static inline Rcpp::List sobol_fast_calc_list(unsigned long long  i, unsigned int dim, unsigned int scramble) {
+static inline Rcpp::List sobol_fast_calc_list(uint64_t  i, unsigned int dim, unsigned int scramble) {
   Rcpp::List vals(dim);
   for(unsigned int j = 0; j < dim; j++) {
     vals[j] = spacefillr::sobol_owen_fast(i, j, scramble);
