+++ src/RcppExports.cpp
@@ -6,12 +6,12 @@
 using namespace Rcpp;
 
 // rcpp_generate_sobol_set
-List rcpp_generate_sobol_set(unsigned long long N, unsigned int dim, unsigned int scramble);
+List rcpp_generate_sobol_set(uint64_t N, unsigned int dim, unsigned int scramble);
 RcppExport SEXP _spacefillr_rcpp_generate_sobol_set(SEXP NSEXP, SEXP dimSEXP, SEXP scrambleSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type scramble(scrambleSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_sobol_set(N, dim, scramble));
@@ -19,12 +19,12 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_generate_sobol_owen_set
-List rcpp_generate_sobol_owen_set(unsigned long long N, unsigned int dim, unsigned int scramble);
+List rcpp_generate_sobol_owen_set(uint64_t N, unsigned int dim, unsigned int scramble);
 RcppExport SEXP _spacefillr_rcpp_generate_sobol_owen_set(SEXP NSEXP, SEXP dimSEXP, SEXP scrambleSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type scramble(scrambleSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_sobol_owen_set(N, dim, scramble));
@@ -32,12 +32,12 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_generate_sobol_owen_fast_set
-List rcpp_generate_sobol_owen_fast_set(unsigned long long N, unsigned int dim, unsigned int scramble);
+List rcpp_generate_sobol_owen_fast_set(uint64_t N, unsigned int dim, unsigned int scramble);
 RcppExport SEXP _spacefillr_rcpp_generate_sobol_owen_fast_set(SEXP NSEXP, SEXP dimSEXP, SEXP scrambleSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type scramble(scrambleSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_sobol_owen_fast_set(N, dim, scramble));
@@ -45,12 +45,12 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_generate_sobol_owen_fast_single
-double rcpp_generate_sobol_owen_fast_single(unsigned long long i, unsigned int dim, unsigned int scramble);
+double rcpp_generate_sobol_owen_fast_single(uint64_t i, unsigned int dim, unsigned int scramble);
 RcppExport SEXP _spacefillr_rcpp_generate_sobol_owen_fast_single(SEXP iSEXP, SEXP dimSEXP, SEXP scrambleSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type i(iSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type i(iSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type scramble(scrambleSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_sobol_owen_fast_single(i, dim, scramble));
@@ -58,24 +58,24 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_generate_halton_faure_set
-List rcpp_generate_halton_faure_set(unsigned long long N, unsigned int dim);
+List rcpp_generate_halton_faure_set(uint64_t N, unsigned int dim);
 RcppExport SEXP _spacefillr_rcpp_generate_halton_faure_set(SEXP NSEXP, SEXP dimSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_halton_faure_set(N, dim));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_generate_halton_random_set
-List rcpp_generate_halton_random_set(unsigned long long N, unsigned int dim, unsigned int seed);
+List rcpp_generate_halton_random_set(uint64_t N, unsigned int dim, unsigned int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_halton_random_set(SEXP NSEXP, SEXP dimSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_halton_random_set(N, dim, seed));
@@ -83,24 +83,24 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_generate_halton_faure_single
-double rcpp_generate_halton_faure_single(unsigned long long i, unsigned int dim);
+double rcpp_generate_halton_faure_single(uint64_t i, unsigned int dim);
 RcppExport SEXP _spacefillr_rcpp_generate_halton_faure_single(SEXP iSEXP, SEXP dimSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type i(iSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type i(iSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_halton_faure_single(i, dim));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_generate_halton_random_single
-double rcpp_generate_halton_random_single(unsigned long long i, unsigned int dim, unsigned int seed);
+double rcpp_generate_halton_random_single(uint64_t i, unsigned int dim, unsigned int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_halton_random_single(SEXP iSEXP, SEXP dimSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type i(iSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type i(iSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type dim(dimSEXP);
     Rcpp::traits::input_parameter< unsigned int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_halton_random_single(i, dim, seed));
@@ -108,60 +108,60 @@ BEGIN_RCPP
 END_RCPP
 }
 // rcpp_generate_pj_set
-List rcpp_generate_pj_set(unsigned long long N, int seed);
+List rcpp_generate_pj_set(uint64_t N, int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_pj_set(SEXP NSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_pj_set(N, seed));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_generate_pmj_set
-List rcpp_generate_pmj_set(unsigned long long N, int seed);
+List rcpp_generate_pmj_set(uint64_t N, int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_pmj_set(SEXP NSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_pmj_set(N, seed));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_generate_pmjbn_set
-List rcpp_generate_pmjbn_set(unsigned long long N, int seed);
+List rcpp_generate_pmjbn_set(uint64_t N, int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_pmjbn_set(SEXP NSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_pmjbn_set(N, seed));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_generate_pmj02_set
-List rcpp_generate_pmj02_set(unsigned long long N, int seed);
+List rcpp_generate_pmj02_set(uint64_t N, int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_pmj02_set(SEXP NSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_pmj02_set(N, seed));
     return rcpp_result_gen;
 END_RCPP
 }
 // rcpp_generate_pmj02bn_set
-List rcpp_generate_pmj02bn_set(unsigned long long N, int seed);
+List rcpp_generate_pmj02bn_set(uint64_t N, int seed);
 RcppExport SEXP _spacefillr_rcpp_generate_pmj02bn_set(SEXP NSEXP, SEXP seedSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
-    Rcpp::traits::input_parameter< unsigned long long >::type N(NSEXP);
+    Rcpp::traits::input_parameter< uint64_t >::type N(NSEXP);
     Rcpp::traits::input_parameter< int >::type seed(seedSEXP);
     rcpp_result_gen = Rcpp::wrap(rcpp_generate_pmj02bn_set(N, seed));
     return rcpp_result_gen;
